utils refactored to intUtils and byteUtils
