+++ src/q344/solution1.go
@@ -1,8 +1,10 @@
 package q344
 
+import byteUtils "github.com/pruthvianveshmuga/dsa/utils/byte"
+
 func Solution1(s []byte) {
 	n := len(s)
 	for i := 0; i < n/2; i++ {
-		Swap(&s[i], &s[n-i-1])
+		byteUtils.Swap(&s[i], &s[n-i-1])
 	}
 }
