+++ src/q350/q350_test.go
@@ -4,7 +4,7 @@ import (
 	"strconv"
 	"testing"
 
-	"github.com/pruthvianveshmuga/dsa/utils"
+	intUtils "github.com/pruthvianveshmuga/dsa/utils/int"
 )
 
 type Case struct {
@@ -23,7 +23,7 @@ func TestSolutions(t *testing.T) {
 			}
 			for tcInd, tc := range tcs {
 				t.Run("c"+strconv.Itoa(tcInd+1), func(t *testing.T) {
-					if res := sol(tc.inp[0], tc.inp[1]); !utils.DeepEqualWithoutOrder(res, tc.out) {
+					if res := sol(tc.inp[0], tc.inp[1]); !intUtils.DeepEqualWithoutOrder(res, tc.out) {
 						t.Errorf("sol%v, testCase%v, %v: expected %v, got %v", solInd, tcInd, tc.inp, tc.out, res)
 					}
 				})
