+++ src/q36/solution1.go
@@ -1,6 +1,6 @@
 package q36
 
-import "github.com/pruthvianveshmuga/dsa/utils"
+import byteUtils "github.com/pruthvianveshmuga/dsa/utils/byte"
 
 func isValid(nums []int) bool {
 	hash := map[int]bool{}
@@ -19,7 +19,7 @@ func Solution1(board [][]byte) bool {
 	for i := 0; i < n; i++ {
 		nums := []int{}
 		for j := 0; j < n; j++ {
-			if val := utils.ByteToInt(board[i][j]); val != 0 {
+			if val := byteUtils.ToInt(board[i][j]); val != 0 {
 				nums = append(nums, val)
 			}
 		}
@@ -31,7 +31,7 @@ func Solution1(board [][]byte) bool {
 	for j := 0; j < n; j++ {
 		nums := []int{}
 		for i := 0; i < n; i++ {
-			if val := utils.ByteToInt(board[i][j]); val != 0 {
+			if val := byteUtils.ToInt(board[i][j]); val != 0 {
 				nums = append(nums, val)
 			}
 		}
@@ -46,7 +46,7 @@ func Solution1(board [][]byte) bool {
 			nums := []int{}
 			for i := 0; i < w; i++ {
 				for j := 0; j < w; j++ {
-					if val := utils.ByteToInt(board[(x*w)+i][(y*w)+j]); val != 0 {
+					if val := byteUtils.ToInt(board[(x*w)+i][(y*w)+j]); val != 0 {
 						nums = append(nums, val)
 					}
 				}
