+++ src/q48/solution1.go
@@ -1,12 +1,12 @@
 package q48
 
-import "github.com/pruthvianveshmuga/dsa/utils"
+import intUtils "github.com/pruthvianveshmuga/dsa/utils/int"
 
 func flipDiagonally(matrix [][]int) {
 	n := len(matrix)
 	for i := 0; i < n; i++ {
 		for j := 0; j < n-i-1; j++ {
-			utils.Swap(&matrix[i][j], &matrix[n-j-1][n-i-1])
+			intUtils.Swap(&matrix[i][j], &matrix[n-j-1][n-i-1])
 		}
 	}
 }
@@ -15,7 +15,7 @@ func flipVertically(matrix [][]int) {
 	n := len(matrix)
 	for i := 0; i < n/2; i++ {
 		for j := 0; j < n; j++ {
-			utils.Swap(&matrix[i][j], &matrix[n-i-1][j])
+			intUtils.Swap(&matrix[i][j], &matrix[n-i-1][j])
 		}
 	}
 }
